booking-service/requirements.txt
fastapi
uvicorn
sqlalchemy
databases[Mysql]
mysql-connector-python

#for loading environment variables  #new
python-dotenv
httpx #for restapi call